azure_function_fastapi_replacement/requirements.txt
azure-functions
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx>=0.24.0

